Change m_distance return type to double

+++ libgeda/include/libgeda/prototype.h
@@ -80,7 +80,7 @@ G_BEGIN_DECLS
          void    libgeda_release                 (void);
 
 /* m_basic.c */
-          int    m_distance                      (int x1, int y1, int x2, int y2);
+       double    m_distance                      (int x1, int y1, int x2, int y2);
          void    m_papersize_to_world            (int width, int height, int border, int *right, int *bottom);
           int    m_random_number                 (int min_num, int max_num);
          void    m_rotate_point                  (int x, int y, int angle, int *newx, int *newy);
