+++ libgeda/src/math/m_basic.c
@@ -39,7 +39,7 @@
  *
  *  \return the distance
  */
-int m_distance(int x1, int y1, int x2, int y2)
+double m_distance(int x1, int y1, int x2, int y2)
 {
   return sqrt(pow(x1-x2,2)+pow(y1-y2,2));
 }
